feat(l10n): Add localization support
- Add Italian localization file
- Update English localization file
- Update package.json with localization properties
- Update tsconfig.json to support resolving JSON modules
